Added the prototype game.
Before running the game-prototype.sh script, make sure you've already
run the build-prototype.sh script so the database file is hanging
around.

To run this script call ./game-prototype.sh with the argument of the
page that you want to read. To begin, simply open at page 1 with
"./game-prototype.sh 1" and follow on from there.